requirements.txt
flask
python-dotenv
customtkinter
requests
pymysql